+++ wallet/bindings/nodejs/examples/1-create-account.js
@@ -1,8 +1,8 @@
 /**
  * This example creates a new database and account
  */
-
-require('dotenv').config({ path: '../.env' });
+const path = require('path')
+require('dotenv').config({ path: path.resolve(__dirname, '.env') });
 const { AccountManager, CoinType } = require('@iota/wallet');
 
 async function run() {