+++ wallet/bindings/nodejs/lib/AccountManager.ts
@@ -288,13 +288,20 @@ export class AccountManager {
      * Replaces client_options, coin_type, secret_manager and accounts. Returns an error if accounts were already created
      * If Stronghold is used as secret_manager, the existing Stronghold file will be overwritten. If a mnemonic was
      * stored, it will be gone.
+     * if ignore_if_coin_type_mismatch is provided client options will not be restored
+     * if ignore_if_coin_type_mismatch == true, client options coin type and accounts will not be restored if the cointype doesn't match
      */
-    async restoreBackup(source: string, password: string): Promise<void> {
+    async restoreBackup(
+        source: string,
+        password: string,
+        ignoreIfCoinTypeMismatch?: boolean,
+    ): Promise<void> {
         await this.messageHandler.sendMessage({
             cmd: 'restoreBackup',
             payload: {
                 source,
                 password,
+                ignoreIfCoinTypeMismatch,
             },
         });
     }